% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affine.R
\name{Affine}
\alias{Affine}
\title{R6 class representing an affine transformation.}
\description{
An affine transformation is given by an invertible 2x2 matrix
(a linear transformation) and a vector (the "intercept").
}
\examples{

## ------------------------------------------------
## Method `Affine$print`
## ------------------------------------------------

Affine$new(rbind(c(3.5,2),c(0,4)), c(-1, 1.25))
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{A}}{get or set the matrix \code{A}}

\item{\code{b}}{get or set the vector \code{b}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Affine$new()}}
\item \href{#method-print}{\code{Affine$print()}}
\item \href{#method-get3x3matrix}{\code{Affine$get3x3matrix()}}
\item \href{#method-inverse}{\code{Affine$inverse()}}
\item \href{#method-compose}{\code{Affine$compose()}}
\item \href{#method-transform}{\code{Affine$transform()}}
\item \href{#method-transformLine}{\code{Affine$transformLine()}}
\item \href{#method-transformEllipse}{\code{Affine$transformEllipse()}}
\item \href{#method-clone}{\code{Affine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Affine} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$new(A, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{A}}{the 2x2 invertible matrix of the affine transformation}

\item{\code{b}}{the shift vector of the affine transformation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Affine} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of an \code{Affine} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Affine$new(rbind(c(3.5,2),c(0,4)), c(-1, 1.25))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get3x3matrix"></a>}}
\subsection{Method \code{get3x3matrix()}}{
The 3x3 matrix representing the affine transformation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$get3x3matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverse"></a>}}
\subsection{Method \code{inverse()}}{
The inverse affine transformation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$inverse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compose"></a>}}
\subsection{Method \code{compose()}}{
Compose the reference affine transformation with another
affine transformation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$compose(transfo, left = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transfo}}{an \code{Affine} object}

\item{\code{left}}{logical, whether to compose at left or at right (i.e.
returns \code{f1 o f0} or \code{f0 o f1})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{Affine} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
Transform a point by the reference affine transformation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$transform(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{a point}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transformLine"></a>}}
\subsection{Method \code{transformLine()}}{
Transform a line by the reference affine transformation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$transformLine(line)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{line}}{a \code{Line} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transformEllipse"></a>}}
\subsection{Method \code{transformEllipse()}}{
Transform an ellipse by the reference affine transformation.
The result is an ellipse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$transformEllipse(ell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ell}}{an \code{Ellipse} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{Ellipse} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Affine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
