% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affine.R
\name{AffineMappingEllipse2Ellipse}
\alias{AffineMappingEllipse2Ellipse}
\title{Affine transformation mapping a given ellipse to a given ellipse}
\usage{
AffineMappingEllipse2Ellipse(ell1, ell2)
}
\arguments{
\item{ell1, ell2}{\code{Ellipse} or \code{Circle} objects}
}
\value{
An \code{Affine} object.
}
\description{
Return the affine transformation which transforms
\code{ell1} to \code{ell2}.
}
\examples{
ell1 <- Ellipse$new(c(1,1), 5, 1, 30)
( ell2 <- Ellipse$new(c(4,-1), 3, 2, 50) )
f <- AffineMappingEllipse2Ellipse(ell1, ell2)
f$transformEllipse(ell1)
}
