% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affine.R
\name{AffineMappingThreePoints}
\alias{AffineMappingThreePoints}
\title{Affine transformation mapping three given points to three given points}
\usage{
AffineMappingThreePoints(P1, P2, P3, Q1, Q2, Q3)
}
\arguments{
\item{P1, P2, P3}{three non-collinear points}

\item{Q1, Q2, Q3}{three non-collinear points}
}
\value{
An \code{Affine} object.
}
\description{
Return the affine transformation which sends
\code{P1} to \code{Q1}, \code{P2} to \code{Q2} and \code{P3} to \code{Q3}.
}
