% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circle.R
\name{Circle}
\alias{Circle}
\title{R6 class representing a circle}
\description{
A circle is given by a center and a radius,
named \code{center} and \code{radius}.
}
\examples{

## ------------------------------------------------
## Method `Circle$new`
## ------------------------------------------------

circ <- Circle$new(c(1,1), 1)
circ
circ$center
circ$center <- c(0,0)
circ

## ------------------------------------------------
## Method `Circle$print`
## ------------------------------------------------

Circle$new(c(0,0), 2)
}
\seealso{
\code{\link{radicalCenter}} for the radical center of three circles.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{center}}{get or set the center}

\item{\code{radius}}{get or set the radius}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Circle$new()}}
\item \href{#method-print}{\code{Circle$print()}}
\item \href{#method-isEqual}{\code{Circle$isEqual()}}
\item \href{#method-power}{\code{Circle$power()}}
\item \href{#method-radicalCenter}{\code{Circle$radicalCenter()}}
\item \href{#method-radicalAxis}{\code{Circle$radicalAxis()}}
\item \href{#method-rotate}{\code{Circle$rotate()}}
\item \href{#method-translate}{\code{Circle$translate()}}
\item \href{#method-invert}{\code{Circle$invert()}}
\item \href{#method-clone}{\code{Circle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Circle} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$new(center, radius)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{the center}

\item{\code{radius}}{the radius}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Circle} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{circ <- Circle$new(c(1,1), 1)
circ
circ$center
circ$center <- c(0,0)
circ
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of a circle object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Circle$new(c(0,0), 2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEqual"></a>}}
\subsection{Method \code{isEqual()}}{
Check whether the circle equals another circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$isEqual(circ)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ}}{a \code{Circle object}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-power"></a>}}
\subsection{Method \code{power()}}{
Power of a point with respect to the circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$power(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A number.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radicalCenter"></a>}}
\subsection{Method \code{radicalCenter()}}{
Radical center of two circles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$radicalCenter(circ2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ2}}{a \code{Circle} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radicalAxis"></a>}}
\subsection{Method \code{radicalAxis()}}{
Radical axis of two circles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$radicalAxis(circ2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ2}}{a \code{Circle} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate"></a>}}
\subsection{Method \code{rotate()}}{
Rotate the circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$rotate(alpha, O, degrees = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{angle of rotation}

\item{\code{O}}{center of rotation}

\item{\code{degrees}}{logical, whether \code{alpha} is given in degrees}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\subsection{Method \code{translate()}}{
Translate the circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$translate(v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v}}{the vector of translation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invert"></a>}}
\subsection{Method \code{invert()}}{
Invert the circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$invert(inversion)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inversion}}{an \code{Inversion} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object or a \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
