% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circle.R
\name{CircleAB}
\alias{CircleAB}
\title{Circle given by a diameter}
\usage{
CircleAB(A, B)
}
\arguments{
\item{A, B}{the endpoints of the diameter}
}
\value{
A \code{Circle} object.
}
\description{
Return the circle given by a diameter
}
