% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circle.R
\name{CircleOA}
\alias{CircleOA}
\title{Circle given by its center and a point}
\usage{
CircleOA(O, A)
}
\arguments{
\item{O}{the center of the circle}

\item{A}{a point of the circle}
}
\value{
A \code{Circle} object.
}
\description{
Return the circle given by its center and a point it
passes through.
}
