% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\alias{draw.Triangle}
\alias{draw.Circle}
\alias{draw.Line}
\title{Draw a geometric object}
\usage{
draw(x, ...)

\method{draw}{Triangle}(x, ...)

\method{draw}{Circle}(x, ...)

\method{draw}{Line}(x, ...)
}
\arguments{
\item{x}{geometric object (\code{Triangle}, \code{Circle} or \code{Line})}

\item{...}{arguments passed to \code{\link{lines}} for a \code{Triangle}
object, to \code{\link[plotrix]{draw.circle}} for a \code{Circle} object,
general graphical parameters for a \code{Line} object}
}
\description{
Draw a geometric object on the current plot.
}
\examples{
# open new plot window
plot(0, 0, type="n", asp = 1, xlim = c(0,2.5), ylim = c(0,2.5),
     xlab = NA, ylab = NA)
grid()
# draw a triangle
t <- Triangle$new(c(0,0), c(1,0), c(0.5,sqrt(3)/2))
draw(t, col = "blue", lwd = 2)
draw(t$rotate(90, t$C), col = "green", lwd = 2)
# draw a line
l <- Line$new(c(1,1), c(1.5,1.5), FALSE, TRUE)
draw(l, col = "red", lwd = 2)
perp <- l$perpendicular(c(2,1))
draw(perp, col = "yellow", lwd = 2)
}
