% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipse.R
\name{EllipseFromFivePoints}
\alias{EllipseFromFivePoints}
\title{Ellipse from five points}
\usage{
EllipseFromFivePoints(P1, P2, P3, P4, P5)
}
\arguments{
\item{P1, P2, P3, P4, P5}{the five points}
}
\value{
An \code{Ellipse} object.
}
\description{
Return an ellipse from five given points on this ellipse.
}
\examples{
ell <- Ellipse$new(c(2,3), 5, 4, 30)
set.seed(666)
pts <- ell$randomPoints(5, "on")
ell2 <- EllipseFromFivePoints(pts[1,],pts[2,],pts[3,],pts[4,],pts[5,])
ell$isEqual(ell2)
}
