% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipse.R
\name{EllipseFromFociAndOnePoint}
\alias{EllipseFromFociAndOnePoint}
\title{Ellipse from foci and one point}
\usage{
EllipseFromFociAndOnePoint(F1, F2, P)
}
\arguments{
\item{F1, F2}{points, the foci}

\item{P}{a point on the boundary of the ellipse}
}
\value{
An \code{Ellipse} object.
}
\description{
Derive the ellipse with given foci and one point on the boundary.
}
