% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipse.R
\name{EllipseFromThreeBoundaryPoints}
\alias{EllipseFromThreeBoundaryPoints}
\title{Smallest ellipse that passes through three boundary points}
\usage{
EllipseFromThreeBoundaryPoints(P1, P2, P3)
}
\arguments{
\item{P1, P2, P3}{three non-collinear points}
}
\value{
An \code{Ellipse} object.
}
\description{
Returns the smallest area ellipse which passes through
  three given boundary points.
}
\examples{
P1 <- c(-1,0); P2 <- c(0, 2); P3 <- c(3,0)
ell <- EllipseFromThreeBoundaryPoints(P1, P2, P3)
ell$includes(P1); ell$includes(P2); ell$includes(P3)
}
