% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipse.R
\name{GaussianEllipse}
\alias{GaussianEllipse}
\title{Gaussian ellipse}
\usage{
GaussianEllipse(mean, Sigma, p)
}
\arguments{
\item{mean}{numeric vector of length 2, the mean of the bivariate Gaussian
distribution; this is the center of the ellipse}

\item{Sigma}{covariance matrix of the bivariate Gaussian distribution}

\item{p}{desired probability level, a number between 0 and 1 (strictly)}
}
\value{
An \code{Ellipse} object.
}
\description{
Return the ellipse equal to the highest \emph{pdf} region of
a bivariate Gaussian distribution with a given probability.
}
