% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperbola.R
\name{HyperbolaFromEquation}
\alias{HyperbolaFromEquation}
\title{Hyperbola object from the hyperbola equation.}
\usage{
HyperbolaFromEquation(eq)
}
\arguments{
\item{eq}{named vector or list of the six parameters \code{Axx}, \code{Axy},
\code{Ayy}, \code{Bx}, \code{By}, \code{C}}
}
\value{
A \code{Hyperbola} object.
}
\description{
Create the \code{Hyperbola} object representing the hyperbola
  with the given implicit equation.
}
