% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersections.R
\name{intersectionCircleLine}
\alias{intersectionCircleLine}
\title{Intersection of a circle and a line}
\usage{
intersectionCircleLine(circ, line)
}
\arguments{
\item{circ}{a \code{Circle} object}

\item{line}{a \code{Line} object}
}
\value{
\code{NULL} if there is no intersection,
a point if the line is tangent to the circle,
a list of two points if the circle and the line meet at
two points.
}
\description{
Return the intersection of a circle and a line.
}
