% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersections.R
\name{intersectionLineLine}
\alias{intersectionLineLine}
\title{Intersection of two lines}
\usage{
intersectionLineLine(line1, line2, strict = FALSE)
}
\arguments{
\item{line1, line2}{two \code{Line} objects}

\item{strict}{logical, whether to take into account the extensions of the
lines (\code{extendA} and \code{extendB})}
}
\value{
If \code{strict = FALSE} this returns either a point, or \code{NULL}
if the lines are parallel, or a bi-infinite line if the two lines coincide.
If \code{strict = TRUE}, this can also return a half-infinite line or
a segment.
}
\description{
Return the intersection of two lines.
}
