% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{inversionFixingThreeCircles}
\alias{inversionFixingThreeCircles}
\title{Inversion fixing three circles}
\usage{
inversionFixingThreeCircles(circ1, circ2, circ3)
}
\arguments{
\item{circ1, circ2, circ3}{\code{Circle} objects}
}
\value{
An \code{Inversion} object, which lets each of \code{circ1},
\code{circ2} and \code{circ3} invariant.
}
\description{
Return the inversion which lets invariant three given circles.
}
