% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{inversionFixingTwoCircles}
\alias{inversionFixingTwoCircles}
\title{Inversion fixing two circles}
\usage{
inversionFixingTwoCircles(circ1, circ2)
}
\arguments{
\item{circ1, circ2}{\code{Circle} objects}
}
\value{
An \code{Inversion} object, which maps \code{circ1} to \code{circ1}
and \code{circ2} to \code{circ2}.
}
\description{
Return the inversion which lets invariant two given circles.
}
