% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{inversionFromCircle}
\alias{inversionFromCircle}
\title{Inversion on a circle}
\usage{
inversionFromCircle(circ)
}
\arguments{
\item{circ}{a \code{Circle} object}
}
\value{
An \code{Inversion} object
}
\description{
Return the inversion on a given circle.
}
