% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{inversionKeepingCircle}
\alias{inversionKeepingCircle}
\title{Inversion keeping a circle unchanged}
\usage{
inversionKeepingCircle(pole, circ)
}
\arguments{
\item{pole}{inversion pole, a point}

\item{circ}{a \code{Circle} object}
}
\value{
An \code{Inversion} object.
}
\description{
Return an inversion with a given pole which keeps a given
circle unchanged.
}
\examples{
circ <- Circle$new(c(4,3), 2)
iota <- inversionKeepingCircle(c(1,2), circ)
iota$transformCircle(circ)
}
