% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{inversionMappingCircle2Circle}
\alias{inversionMappingCircle2Circle}
\title{Inversion mapping a circle to a circle}
\usage{
inversionMappingCircle2Circle(circ1, circ2, positive = FALSE)
}
\arguments{
\item{circ1, circ2}{\code{Circle} objects}

\item{positive}{logical, whether the sign of the desired inversion power
must be positive or negative}
}
\value{
An \code{Inversion} object, which maps \code{circ1} to \code{circ2}.
}
\description{
Return the inversion which maps a given circle to another given
circle.
}
