% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{inversionSwappingTwoCircles}
\alias{inversionSwappingTwoCircles}
\title{Inversion swapping two circles}
\usage{
inversionSwappingTwoCircles(circ1, circ2, positive = TRUE)
}
\arguments{
\item{circ1, circ2}{\code{Circle} objects}

\item{positive}{logical, whether the sign of the desired inversion power
must be positive or negative}
}
\value{
An \code{Inversion} object, which maps \code{circ1} to \code{circ2}
and \code{circ2} to \code{circ1}.
}
\description{
Return the inversion which swaps two given circles.
}
