% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineFrom.R
\name{LineFromEquation}
\alias{LineFromEquation}
\title{Line from general equation}
\usage{
LineFromEquation(a, b, c)
}
\arguments{
\item{a, b, c}{the parameters of the equation; \code{a} and \code{b} cannot
be both zero}
}
\value{
A \code{Line} object.
}
\description{
Create a \code{Line} object representing the infinite line
  with given equation \eqn{ax + by + c = 0}.
}
