% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineFrom.R
\name{LineFromInterceptAndSlope}
\alias{LineFromInterceptAndSlope}
\title{Line from intercept and slope}
\usage{
LineFromInterceptAndSlope(a, b)
}
\arguments{
\item{a}{intercept}

\item{b}{slope}
}
\value{
A \code{Line} object.
}
\description{
Create a \code{Line} object representing the infinite line
  with given intercept and given slope
}
