% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mobius.R
\name{MobiusMappingThreePoints}
\alias{MobiusMappingThreePoints}
\title{Möbius transformation mapping three given points to three given points}
\usage{
MobiusMappingThreePoints(P1, P2, P3, Q1, Q2, Q3)
}
\arguments{
\item{P1, P2, P3}{three distinct points, \code{Inf} allowed}

\item{Q1, Q2, Q3}{three distinct points, \code{Inf} allowed}
}
\value{
A \code{Mobius} object.
}
\description{
Return a Möbius transformation which sends
\code{P1} to \code{Q1}, \code{P2} to \code{Q2} and \code{P3} to \code{Q3}.
}
