% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mobius.R
\name{MobiusSwappingTwoPoints}
\alias{MobiusSwappingTwoPoints}
\title{Möbius transformation swapping two given points}
\usage{
MobiusSwappingTwoPoints(A, B)
}
\arguments{
\item{A, B}{two distinct points, \code{Inf} not allowed}
}
\value{
A \code{Mobius} object.
}
\description{
Return a Möbius transformation which sends
  \code{A} to \code{B} and \code{B} to \code{A}.
}
