% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circle.R
\name{soddyCircle}
\alias{soddyCircle}
\title{Inner Soddy circle}
\usage{
soddyCircle(circ1, circ2, circ3)
}
\arguments{
\item{circ1, circ2, circ3}{distinct circles}
}
\value{
A \code{Circle} object.
}
\description{
Inner Soddy circles associated to three circles.
}
