% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangle.R
\name{TriangleThreeLines}
\alias{TriangleThreeLines}
\title{Triangle defined by three lines}
\usage{
TriangleThreeLines(line1, line2, line3)
}
\arguments{
\item{line1, line2, line3}{\code{Line} objects}
}
\value{
A \code{Triangle} object.
}
\description{
Return the triangle formed by three lines.
}
