% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circle.R
\docType{data}
\name{unitCircle}
\alias{unitCircle}
\title{Unit circle}
\format{
An object of class \code{Circle} (inherits from \code{R6}) of length 24.
}
\usage{
unitCircle
}
\description{
Circle centered at the origin with radius 1.
}
\keyword{datasets}
