% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arc.R
\name{Arc}
\alias{Arc}
\title{R6 class representing a circular arc}
\description{
An arc is given by a center, a radius, a starting angle and
an ending angle. They are respectively named \code{center}, \code{radius},
\code{alpha1} and \code{alpha2}.
}
\examples{

## ------------------------------------------------
## Method `Arc$new`
## ------------------------------------------------

arc <- Arc$new(c(1,1), 1, pi/4, pi/2)
arc
arc$center
arc$center <- c(0,0)
arc

## ------------------------------------------------
## Method `Arc$print`
## ------------------------------------------------

Arc$new(c(0,0), 2, pi/4, pi/2)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{center}}{get or set the center}

\item{\code{radius}}{get or set the radius}

\item{\code{alpha1}}{get or set the starting angle}

\item{\code{alpha2}}{get or set the ending angle}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Arc$new()}}
\item \href{#method-print}{\code{Arc$print()}}
\item \href{#method-startingPoint}{\code{Arc$startingPoint()}}
\item \href{#method-endingPoint}{\code{Arc$endingPoint()}}
\item \href{#method-isEqual}{\code{Arc$isEqual()}}
\item \href{#method-complementaryArc}{\code{Arc$complementaryArc()}}
\item \href{#method-path}{\code{Arc$path()}}
\item \href{#method-clone}{\code{Arc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Arc} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$new(center, radius, alpha1, alpha2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{the center}

\item{\code{radius}}{the radius}

\item{\code{alpha1}}{the starting angle}

\item{\code{alpha2}}{the ending angle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Arc} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{arc <- Arc$new(c(1,1), 1, pi/4, pi/2)
arc
arc$center
arc$center <- c(0,0)
arc
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of a circle object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Arc$new(c(0,0), 2, pi/4, pi/2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-startingPoint"></a>}}
\subsection{Method \code{startingPoint()}}{
Starting point of the reference arc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$startingPoint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-endingPoint"></a>}}
\subsection{Method \code{endingPoint()}}{
Ending point of the reference arc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$endingPoint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEqual"></a>}}
\subsection{Method \code{isEqual()}}{
Check whether the reference arc equals another arc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$isEqual(arc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arc}}{an \code{Arc} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-complementaryArc"></a>}}
\subsection{Method \code{complementaryArc()}}{
Complementary arc of the reference arc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$complementaryArc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\subsection{Method \code{path()}}{
Path defining the reference arc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$path(npoints = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{npoints}}{number of points of the path}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with two numeric vectors \code{x} and \code{y} of length \code{npoints}.
See "Filling the lapping area of two circles" in the vignette for
an example.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
