% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circle.R
\name{Circle}
\alias{Circle}
\title{R6 class representing a circle}
\description{
A circle is given by a center and a radius,
named \code{center} and \code{radius}.
}
\examples{

## ------------------------------------------------
## Method `Circle$new`
## ------------------------------------------------

circ <- Circle$new(c(1,1), 1)
circ
circ$center
circ$center <- c(0,0)
circ

## ------------------------------------------------
## Method `Circle$print`
## ------------------------------------------------

Circle$new(c(0,0), 2)

## ------------------------------------------------
## Method `Circle$orthogonalThroughTwoPointsOnCircle`
## ------------------------------------------------

# hyperbolic triangle
circ <- Circle$new(c(5,5), 3)
arc1 <- circ$orthogonalThroughTwoPointsOnCircle(0, 2*pi/3, arc = TRUE)
arc2 <- circ$orthogonalThroughTwoPointsOnCircle(2*pi/3, 4*pi/3, arc = TRUE)
arc3 <- circ$orthogonalThroughTwoPointsOnCircle(4*pi/3, 0, arc = TRUE)
opar <- par(mar = c(0,0,0,0))
plot(0, 0, type = "n", asp = 1, xlim = c(2,8), ylim = c(2,8))
draw(circ)
draw(arc1, col = "red", lwd = 2)
draw(arc2, col = "green", lwd = 2)
draw(arc3, col = "blue", lwd = 2)
par(opar)
}
\seealso{
\code{\link{radicalCenter}} for the radical center of three circles.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{center}}{get or set the center}

\item{\code{radius}}{get or set the radius}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Circle$new()}}
\item \href{#method-print}{\code{Circle$print()}}
\item \href{#method-pointFromAngle}{\code{Circle$pointFromAngle()}}
\item \href{#method-isEqual}{\code{Circle$isEqual()}}
\item \href{#method-includes}{\code{Circle$includes()}}
\item \href{#method-orthogonalThroughTwoPointsOnCircle}{\code{Circle$orthogonalThroughTwoPointsOnCircle()}}
\item \href{#method-orthogonalThroughTwoPointsWithinCircle}{\code{Circle$orthogonalThroughTwoPointsWithinCircle()}}
\item \href{#method-power}{\code{Circle$power()}}
\item \href{#method-radicalCenter}{\code{Circle$radicalCenter()}}
\item \href{#method-radicalAxis}{\code{Circle$radicalAxis()}}
\item \href{#method-rotate}{\code{Circle$rotate()}}
\item \href{#method-translate}{\code{Circle$translate()}}
\item \href{#method-invert}{\code{Circle$invert()}}
\item \href{#method-clone}{\code{Circle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Circle} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$new(center, radius)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{the center}

\item{\code{radius}}{the radius}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Circle} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{circ <- Circle$new(c(1,1), 1)
circ
circ$center
circ$center <- c(0,0)
circ
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of a circle object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Circle$new(c(0,0), 2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pointFromAngle"></a>}}
\subsection{Method \code{pointFromAngle()}}{
Get a point on the reference circle from its polar angle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$pointFromAngle(alpha, degrees = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{a number, the angle}

\item{\code{degrees}}{logical, whether \code{alpha} is given in degrees}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The point on the circle with polar angle \code{alpha}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEqual"></a>}}
\subsection{Method \code{isEqual()}}{
Check whether the reference circle equals another circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$isEqual(circ)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ}}{a \code{Circle} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-includes"></a>}}
\subsection{Method \code{includes()}}{
Check whether a point belongs to the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$includes(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{a point}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orthogonalThroughTwoPointsOnCircle"></a>}}
\subsection{Method \code{orthogonalThroughTwoPointsOnCircle()}}{
Orthogonal circle passing through two points on the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$orthogonalThroughTwoPointsOnCircle(alpha1, alpha2, arc = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha1, alpha2}}{two angles defining two points on the reference circle}

\item{\code{arc}}{logical, whether to return only the arc at the interior of the
reference circle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object if \code{arc=FALSE}, an \code{Arc} object
if \code{arc=TRUE}, or a \code{Line} object: the diameter
of the reference circle defined by the two points in case when the two
angles differ by \code{pi}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# hyperbolic triangle
circ <- Circle$new(c(5,5), 3)
arc1 <- circ$orthogonalThroughTwoPointsOnCircle(0, 2*pi/3, arc = TRUE)
arc2 <- circ$orthogonalThroughTwoPointsOnCircle(2*pi/3, 4*pi/3, arc = TRUE)
arc3 <- circ$orthogonalThroughTwoPointsOnCircle(4*pi/3, 0, arc = TRUE)
opar <- par(mar = c(0,0,0,0))
plot(0, 0, type = "n", asp = 1, xlim = c(2,8), ylim = c(2,8))
draw(circ)
draw(arc1, col = "red", lwd = 2)
draw(arc2, col = "green", lwd = 2)
draw(arc3, col = "blue", lwd = 2)
par(opar)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orthogonalThroughTwoPointsWithinCircle"></a>}}
\subsection{Method \code{orthogonalThroughTwoPointsWithinCircle()}}{
Orthogonal circle passing through two points within the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$orthogonalThroughTwoPointsWithinCircle(P1, P2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{P1, P2}}{two points in the interior of the reference circle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-power"></a>}}
\subsection{Method \code{power()}}{
Power of a point with respect to the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$power(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A number.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radicalCenter"></a>}}
\subsection{Method \code{radicalCenter()}}{
Radical center of two circles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$radicalCenter(circ2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ2}}{a \code{Circle} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radicalAxis"></a>}}
\subsection{Method \code{radicalAxis()}}{
Radical axis of two circles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$radicalAxis(circ2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ2}}{a \code{Circle} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate"></a>}}
\subsection{Method \code{rotate()}}{
Rotate the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$rotate(alpha, O, degrees = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{angle of rotation}

\item{\code{O}}{center of rotation}

\item{\code{degrees}}{logical, whether \code{alpha} is given in degrees}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\subsection{Method \code{translate()}}{
Translate the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$translate(v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v}}{the vector of translation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invert"></a>}}
\subsection{Method \code{invert()}}{
Invert the reference circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$invert(inversion)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inversion}}{an \code{Inversion} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object or a \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Circle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
