% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipse.R
\name{Ellipse}
\alias{Ellipse}
\title{R6 class representing an ellipse}
\description{
An ellipse is given by a center, two radii (\code{rmajor} and \code{rminor}),
and the angle (\code{alpha}) between the major axis and the horizontal direction.
}
\examples{

## ------------------------------------------------
## Method `Ellipse$new`
## ------------------------------------------------

Ellipse$new(c(1,1), 3, 2, 30)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{center}}{get or set the center}

\item{\code{rmajor}}{get or set the major radius of the ellipse}

\item{\code{rminor}}{get or set the minor radius of the ellipse}

\item{\code{alpha}}{get or set the angle of the ellipse}

\item{\code{degrees}}{get or set the \code{degrees} field}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ellipse$new()}}
\item \href{#method-print}{\code{Ellipse$print()}}
\item \href{#method-equation}{\code{Ellipse$equation()}}
\item \href{#method-includes}{\code{Ellipse$includes()}}
\item \href{#method-contains}{\code{Ellipse$contains()}}
\item \href{#method-clone}{\code{Ellipse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Ellipse} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ellipse$new(center, rmajor, rminor, alpha, degrees = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{a point, the center of the rotation}

\item{\code{rmajor}}{positive number, the major radius}

\item{\code{rminor}}{positive number, the minor radius}

\item{\code{alpha}}{a number, the angle between the major axis and the horizontal direction}

\item{\code{degrees}}{logical, whether \code{alpha} is given in degrees}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Ellipse} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Ellipse$new(c(1,1), 3, 2, 30)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of an \code{Ellipse} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ellipse$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-equation"></a>}}
\subsection{Method \code{equation()}}{
The coefficients of the implicit equation of the ellipse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ellipse$equation()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The implicit equation of the ellipse is
\code{Ax² + Bxy + Cy² + Dx + Ey + F = 0}. This method returns
A, B, C, D, E and F.
}

\subsection{Returns}{
A named numeric vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-includes"></a>}}
\subsection{Method \code{includes()}}{
Check whether a point lies on the reference ellipse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ellipse$includes(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{a point}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contains"></a>}}
\subsection{Method \code{contains()}}{
Check whether a point is contained in the reference ellipse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ellipse$contains(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{a point}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ellipse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
