% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inversion.R
\name{Inversion}
\alias{Inversion}
\title{R6 class representing an inversion}
\description{
An inversion is given by a pole (a point) and a power (a number,
possibly negative).
}
\examples{

## ------------------------------------------------
## Method `Inversion$print`
## ------------------------------------------------

Inversion$new(c(0,0), 2)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pole}}{get or set the pole}

\item{\code{power}}{get or set the power}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Inversion$new()}}
\item \href{#method-print}{\code{Inversion$print()}}
\item \href{#method-invert}{\code{Inversion$invert()}}
\item \href{#method-invertCircle}{\code{Inversion$invertCircle()}}
\item \href{#method-invertLine}{\code{Inversion$invertLine()}}
\item \href{#method-clone}{\code{Inversion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Inversion} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inversion$new(pole, power)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pole}}{the pole}

\item{\code{power}}{the power}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Inversion} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of an inversion object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inversion$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Inversion$new(c(0,0), 2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invert"></a>}}
\subsection{Method \code{invert()}}{
Inversion of a point.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inversion$invert(M)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{a point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A point.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invertCircle"></a>}}
\subsection{Method \code{invertCircle()}}{
Inversion of a circle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inversion$invertCircle(circ)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{circ}}{a \code{Circle} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object or a \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invertLine"></a>}}
\subsection{Method \code{invertLine()}}{
Inversion of a line.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inversion$invertLine(line)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{line}}{a \code{Line} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Circle} object or a \code{Line} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inversion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
