% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.R
\name{Line}
\alias{Line}
\title{R6 class representing a line}
\description{
A line is given by two distinct points,
named \code{A} and \code{B}, and two logical values \code{extendA}
and \code{extendB}, indicating whether the line must be extended
beyond \code{A} and \code{B} respectively.
}
\examples{

## ------------------------------------------------
## Method `Line$new`
## ------------------------------------------------

l <- Line$new(c(1,1), c(1.5,1.5), FALSE, TRUE)
l
l$A
l$A <- c(0,0)
l

## ------------------------------------------------
## Method `Line$print`
## ------------------------------------------------

Line$new(c(0,0), c(1,0), FALSE, TRUE)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{A}}{get or set the point A}

\item{\code{B}}{get or set the point B}

\item{\code{extendA}}{get or set \code{extendA}}

\item{\code{extendB}}{get or set \code{extendB}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Line$new()}}
\item \href{#method-print}{\code{Line$print()}}
\item \href{#method-clone}{\code{Line$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{Line} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Line$new(A, B, extendA, extendB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{A, B}}{points}

\item{\code{extendA, extendB}}{logical values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Line} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{l <- Line$new(c(1,1), c(1.5,1.5), FALSE, TRUE)
l
l$A
l$A <- c(0,0)
l
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Show instance of a line object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Line$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Line$new(c(0,0), c(1,0), FALSE, TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Line$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
